backend/requirements.txt
langchain==1.0.0
langchain-core==1.0.0
langsmith>=0.4.19,<0.4.35
langchain-community==0.4
langchain-groq==1.0.0
faiss-cpu==1.12.0
sentence-transformers==3.0.1
openai==1.40.6
tiktoken==0.7.0
pypdf==4.2.0
python-dotenv==1.0.1
requests
numpy==1.26.4
pandas==2.2.2
packaging==24.1
typing-extensions==4.12.2
groq>=0.4.0
python-multipart>=0.0.6

